python_backend/requirements.txt
# AlumConnect Python Backend Dependencies

# Web Framework
fastapi==0.115.0
uvicorn[standard]==0.30.0

# Google AI
google-genai==1.14.0

# LangChain
langchain==0.3.25
langchain-google-genai==2.1.4
langchain-community==0.3.24

# ML & Data
scikit-learn==1.6.1
numpy==2.2.5
pandas==2.2.3

# Firebase
firebase-admin==6.7.0

# PDF Parsing
PyPDF2==3.0.1
python-docx==1.1.2

# Utilities
python-dotenv==1.1.0
pydantic==2.11.3
httpx==0.28.1
python-multipart==0.0.20

# Testing
pytest==8.3.4
pytest-asyncio==0.24.0
pytest-cov==6.0.0
httpx==0.28.1
mutmut==3.2.1
